Fix lets
Previously we we implemented lets by adding a variable to the context
and then setting the variable's projection equal to the term of the
body. This is wrong since it will set these terms equal in all equal
contexts, not just the current one. So change to keep track of defined
terms in the context in addition to allowing extending the contexts by
new types. Defining such a term in a context does not affect the type of
the context.