Be a tryhard and make process_root return io::Result
Probably not worth it; at most, it might sense to return a Result<(),
Box<dyn Error>> perhaps. As it is, we panic when an eqlog file doesn't
type check, and if we want return it we'd have to wrap it in an
io::Error. This would make the error message less nice if in build.rs we
just call unwrap().