Try to depend on eqlog-eqlog from eqlog
This doesn't quite work yet, since now we're using eqlog and
eqlog-runtime in two versions each: Once the published version on
crates.io, and once the local version. The workaround is to rename the
crates.io version. However, eqlog assumes that the runtime crate is
called eqlog-runtime. That assumption is now broken, so that'll have to be fixed first.