Revert runtime crate config
This reverts commit 34525ceb6cae73fa148aa66ed9f3986e2a157aea.
This reverts commit 10f457f22236d59e98fee0c7ae8a5e7d2cdbcc98.

Turns out this wasn't actually needed for eqlog-eqlog to work. The only
thing that's not working is `cargo build -p eqlog`, since `eqlog` is the
name both of a local crate and a remote crate. Instead, `-p eqlog` has
to be replaced with something like `-p file://...`. But that's necessary
even when the eqlog dependency in eqlog-eqlog is renamed to something
else, so renaming is not useful after all.