Revert test migration to standalone crate
This is a revert of the following commits:
- 8eb6e0d342e345ab614387a8a424136db8156e77
- 06002e8887014bace5ee6085f698983bf4e0031a
- 2c10bb8beb713f4978becc135c216009d8d05fdf

For test cases where we expect eqlog compilation to succeed (currently
these are the only tests) standalone crates for each test case are
overly granular. I did the migration anyway because I want to add tests
where compilation does not succeed. If we ran such tests via cargo and
build.rs scripts, then we would need a separate crate for each test
case. But we can just call e.g. process_root directly in a normal test
case, and then check that it paniced/crashed. Calling process_root
directly is worse in terms of end-to-end coverage, but the simplicity
compared to orchestrating tests that run it via cargo and a build.rs
file probably outweighs this.