Migrate symbol lookup checks to use eqlog
This made it necessary to change the type lookup check, since the
existing one will also result in an undetermined type error.
Notice that we sort errors according to their *end* location.
Since the undetermined type error for an if atom `x : Ty` is reported
for the term `x` but the type lookup error for the whole atom `x: Ty`,
the former takes precedence.