Replace Hash{Set,Map} by BTree variants
This is to make the output of eqlog deterministic. HashMap and HashSet
use a different seed that differs each time the executable is run, so
their usage results in non-determinism. This could also have been
resolved by fixing this seed, but the btree solution is simpler and
probably also easier to debug.