Fix ordering in seminaive lowering
If eqlog is itself compiled with this fix, this brings compilation of
examples/inference down from over 30 seconds to under 10 milliseconds
(!).

The problem was that we were sometimes scanning full relations in inner
loops, specifically for the long rules in eqlog.eqlog concerning grouped
chains. For some rules full scanning in inner loops cannot be avoided,
but before this commit we didn't even try. Now we're always selecting
the next inner loop in such a way that we maximize the number of already
fixed arguments in the relation, which can drastically reduces the
number of iterations for that inner loop.