Bump eqlog dep version of eqlog-eqlog to 0.6
To answer the question in the previous commit: Yes, cargo appears to
dislike two versions of the same package which differ only in a patch
version. This means that whenever the eqlog version that eqlog-eqlog depends on should be bumped, the minor (not patch) version of eqlog should be bumped before.