Fix usage of the rebuild feature
Turns out that features for a crate are actually not unified over a
workspace. This meant that running `cargo build -p eqlog` would build
eqlog without the "rebuild" feature, whereas some crates that depended
on eqlog with "rebuild" turned on would enable it.

This commit changes usage of the "rebuild" features as as follows:

- If the eqlog create feature "rebuild" is turned on (it's not by
  default), then also the "rebuild" feature of eqlog-eqlog is turned on
  (it's not by default).
- All local crates that depend on eqlog turn on the "rebuild" feature on
  it. This includes the eqlog-test-* crates.

This means that building eqlog or eqlog-eqlog in isolation (e.g. `cargo
build -p eqlog`) will actually build those crates with "rebuild"
*disabled*. For an actual rebuild, we have to build one of the crates
that depend on the eqlog crate, for example eqlog-test-compile.