Fix bug in b9b1fd938: Use actual model insert fn
... and not the insert fn of the table inside the model. This also fixes
the mystery noticed in 5727c46eaf1, why saturating_add/sub are necessary
for the tests to pass: Indeed, we weren't updating weights as we should.
With this commit, the tests pass also if we use the unsaturating
operations. It still makes sense to keep them saturating though, becuase
it's probably not much of a difference (if at all) in terms of runtime,
and it's more correct the way it is in case we actually overflow.