Fix match statement desugaring
Previously, we desugared to an equality `if <var> = <pattern>` for each
match case, where `var` is a new variable. I'm not sure what the
intention here was, but this was equivalent to just `if <pattern>!`,
which just matches the pattern with anything in the model, not
necessarily the discriminee. Now we're desugaring to `if <discriminee> =
<pattern>`.