Fix flatten
Now that we have branching, just pushing new if statements onto function
that matches the domain is a bug: This makes the function match more
than just the domain, and in case there's a different morphism with the
same domain (i.e., a branch), we'd again match this data from the first
branch in the second branch.

TODO: The same bug is also present when flattening non surjective
functions. Note that a non-surjective then statement functions as an if
statement matching arbitrary (not necessarily fresh) data in case the
potentially missing element is actually there.