Fix perf problem by splitting rule
This rule was matching too much data, and as a result took very long. By
splitting the hypotheses of that rule up into a few smaller ones,
execution of the rule became vastly faster: From dominating the
execution time to an irrelevant fraction. Perhaps something that should
be done automatically?