Bump version to 0.8.0
The source, apart from the version numbers, is no difference to 0.7.3, but with 0.7.3 I get cargo conflicting dependencies when trying to publish eqlog:

```
> cargo publish -p eqlog
error: failed to prepare local package for uploading

Caused by:
  failed to select a version for `eqlog-eqlog`.
      ... required by package `eqlog v0.7.2`
      ... which satisfies dependency `eqlog = "=0.7.2"` of package `eqlog-eqlog v0.7.3`
      ... which satisfies dependency `eqlog-eqlog = "=0.7.3"` of package `eqlog v0.7.3 (/home/mbid/projects/eqlog/eqlog)`
  versions that meet the requirements `=0.7.2` are: 0.7.2

  all possible versions conflict with previously selected packages.

    previously selected package `eqlog-eqlog v0.7.3`
      ... which satisfies dependency `eqlog-eqlog = "=0.7.3"` of package `eqlog v0.7.3 (/home/mbid/projects/eqlog/eqlog)`

  failed to select a version for `eqlog-eqlog` which could resolve this conflict
```

I got this after I had successfully published eqlog-runtime and
eqlog-eqlog at 0.7.3. In the past I think increasing to a minor version,
isntead of a patch version, was sufficient to resolve these issues.
Perhaps you can't depend on two different version of the same crate that
differ only in the last digit, the patch? It would be reasonable not to
allow that if you want to adhere to semver.