Get rid of table iterator fns/types
This required migrating also NonSurjThen away from those.
The commit adds an eval fn to the table struct.
This should also be used in the public API method, but isn't at the
moment.
There's also the problem that we're generating all possible index
getter decls in rule modules instead of only those that are actually
used.