Revert "Split up apply tuples fn"
This reverts commit c55217d1a440ee5b56919f8f98def639f2beaf40.

The only thing gained here would've been that we don't do something like
this sequence:

1. Insert new tuple x
2. Topologically sort
3. Propagate data along morphisms, potentially remove x again

But this should be a constant or at least log(n) factor somewhere in the
best case. If we instead wanted to insert new tuples intertwined with
propagating data, then we could save something, but this complexity is
not worth what we'd get.